Item 03983: Broadcast last event to in-progress Publishing box - eliminate the spinner
Remove the spinner.

The publishing event log is implemented using a reverse-sorted ordered list for usability and accessibility.
The numbered list is not desired, but provided for accessibility and therefore the numbers are hidden via CSS.

Implement a custom (simple/naive) filter for event-type-specific presentation.
Render the event timestamp for improved user experience.

Ensure that when publishing is off and a document update is received that the events log is cleared.

Help prevent fast clicking on publish submit button from triggering multiple publishes.
(This is still blocked by the service-end, but doing this at least prevents an annoying publish denied error.)