Item D-01293: Can navigate to annotate for item that's already published
Redirect to appropriate "/view" path when in an invalid status for annotation.

The console will show unsubscribe() calls after the subscribe() calls after this redirect.
These calls make it appear as if the listeners are unsubscribed when they are not.
It seems that the order of the calls can be safely ignored.