Fix merge regression between publishing log and annotate restriction branches
The controller for the "view" seems to be the same controller for "annotate".
The resulting logic in annotation access control checks prevents the publishing log listeners from being assigned.

It turns out that the path can be changed via assigning `$scope.action`.
The `$location.path()` redirect can be replaced with a `$scope.action`, simplifying the code.

Fix whitespace issues resulting from the merge.