Item B-03986: Create Field Profile modal improvements
Handle alerts in the model.
Clear the alerts on close.
Provide a cancel button.
Treat the cancel button and X (close) button as the same.

Validation has not been implemented at this time because there appears to be a lot of significant design changes that would be necessary to utilize the weaver validation.

While I would prefer to have the buttons in the button section of the model and the messages across the top, these changes were avoided to ensure consistency in existing design.